Base code: Window, app lifecyle, and event loop.
The application begins by creating a Winit
[EventLoop](https://docs.rs/winit/0.25.0/winit/event_loop/struct.EventLoop.html)
to handle events from the operating system and window. The event loop is
used to create a
[Window](https://docs.rs/winit/0.25.0/winit/window/struct.Window.html)
using the window builder API.

Once the window is initialized, we run the application main loop which
polls for events from the event loop and exits when a window close event
is detected.

After the main loop completes, clean up is executed which currently does
nothing since `Window` implements the Drop trait which automatically
cleans up the platform specific window resources.

Note: The event loop nor the window are members of the application
struct. The event loop cannot be a part of the application struct due to
ownership conflicts. There may be workarounds for this but no need to
prematurely optimise. The window is not currently used either.

---

References:
* https://docs.rs/winit/0.25.0/winit/
* https://github.com/unknownue/vulkan-tutorial-rust/blob/master/src/tutorials/00_base_code.rs
* https://github.com/bwasty/vulkan-tutorial-rs/blob/master/src/bin/00_base_code.rs
* https://vulkan-tutorial.com/en/Drawing_a_triangle/Setup/Base_code