Vulkan instance creation
Creating an instance involves a few steps:

Load the Vulkan function pointers

To call Vulkan functions, we need to load the function pointers. This is
handled by the `ash::Entry` type. We create our entry in an unsafe block
which loads the function pointers. The entry is stored on the
application structure so it can be used for Vulkan calls later

Check available extensions

We use VkEnumerateInstanceExtensionProperties to check all of the
extensions that are available to us. We need to use the Surface,
Win32Surface extensions to run on windows. We use the DebugUtils
extension so we can use debugging tools.

Application Info

Application Info defines metadata for the application such as name and
version that the driver can use to make optimisations.

Create Instance

We create an InstanceCreateInfo that references the extensions we want
to load and the application info. This is used as a parameter to the
VkCreateInstance function which we use the entry we created earlier to
call.

Cleanup

We implement the Drop trait for the application so that the Vulkan
instance is destroyed when our application exits. This ensures there are
no resource leakages when the program exits.

Notes:

* Since the Vulkan API uses a c-like interface it expects most strings
  as raw pointers. We make liberal use of the CString crate to define
  and create references to raw string pointers.
* Many of the Vulkan structs have fields we don't care about. Rust
  doesn't allow us to leave members of a struct empty so they must be
  set to null pointers using ptr::null(). Thankfully, ash provides
  builder structs which can incrementally build a vulkan struct and
  hides the need to set null values.
* We don't really need a init and cleanup method. Initialisation is
  handled in the "constructor" method for the application. Cleanup is
  handled by implementing `Drop` for the application so that when it
  falls out of scope the associated resources are released