Debut Utils messenger callback
Sets up the messenger callback for the Debug Utils extensions.

https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#VK_EXT_debug_utils

Ash provides a struct `DebugUtils` which loads the debug utils functions
for us. We define a callback which reads the message severity, message
type, and message callback and create the message handler object
associated with our Instance object using the extension's
`vkCreateDebugUtilsMessengerEXT` function. Additionally we provide the
same messenger as the `pNext` when creating the instance so it can also
debug the instance creation and destruction. The messenger is cleaned up
by the extension as part of the `Drop` implementation of the
application.

Additionally this commit fixes a bug where `Drop` isn't being called at
the end of the program. `winit`'s `EventLoop#run` hijacks the main
thread resulting in resources in the old thread not being dropped. We
create a placeholder function `draw_frame` which we can call from the
event loop callback. This ensures it is dropped when the event loop
exits.

Resources:

- https://vulkan-tutorial.com/Drawing_a_triangle/Setup/Validation_layers
- https://docs.rs/winit/0.24.0/winit/event_loop/struct.EventLoop.html#method.run
- https://github.com/unknownue/vulkan-tutorial-rust/blob/master/src/tutorials/02_validation_layers.rs
- https://github.com/unknownue/vulkan-tutorial-rust/issues/14