Physical device and queue families
Finds the first suitable physical device using
VkEnumeratePhysicalDevices. Only discrete GPUs with support
for geometry shaders are suitable.

Additionally, we read the queue families available on the physical
device using get_physical_device_queue_family_properties. We make sure
that the device has a graphics queue family and store the index so that
we can reference the queue later.

References

- https://vulkan-tutorial.com/Drawing_a_triangle/Setup/Physical_devices_and_queue_families