Logical device and graphics queue
We use the physical device to get a logical device from the VkInstance
using vkCreateDevice. This requires a device_create_info which includes
the device features which you want to use (that we queried from the
physical device e.g. geometry shaders), and an array of queue create
infos that specify the families of the queues that we want to create. In
this instance we only create a graphics queue. The queue object is
created and destroyed automatically with the logical device. However, we
need to get a handle on the queue by using VkGetDeviceQueue witht he
specified queue family and queue index. We store this handle on our
application struct.

We add the destruction of the logical device to the `Drop`
implementation of the application.