Surface creation and presentation queue
We load the required extension to create a surface, then use the window
object from `winit` to create a surface from the native window
handle. Only Win32 surfaces are currently supported. We import the win32
api crate so we can get the native HWND from the winit window and the
ModuleHandle/hinstance from the current process. We then use the
surface_loader (analogous to ash::Entry but for the Win32 Surface
extension) to create the Surface from the HWND and Hinstance.

We do this before picking our physical device so we can make sure that
the physical device provides a queue that supports presenting to our surface. We need to use
the surface loader here so we can use the
get_physical_device_surface_support extension function. We save the
queue index that supports presenting to our queue index families.

We modify the logical device creation to request a presentation queue as
well as a graphics queue. Sometimes these may be the same underlying
queue so we make sure to reduce our queue family indices to only contain
unique values. This ensures we don't request the same queue be created multiple
times.

Once the device is created, we can fetch the presentation queue from the
logical device the same way we do the graphics queue.

Lastly we modify the Drop function to make sure the Surface is cleaned
up. We need to use the surface_loader extension functions here too.

Resources

- https://vulkan-tutorial.com/Drawing_a_triangle/Presentation/Window_surface
- https://docs.microsoft.com/en-us/windows/win32/api/libloaderapi/nf-libloaderapi-getmodulehandlew
- https://docs.microsoft.com/en-us/windows/win32/learnwin32/creating-a-window#creating-the-window