WIP: Swap Chain
Enable the swap chain extension on the logical device if it is present
by adding the extension name to the enabled_extensions of the device
create info.

We also begin determining how to build our swap chain. First we query
the supported features of the swap chain -- capabilities (Swap
extents/resolution of images), surface
formats (colour space and and colour format), and presentation modes
(how images are swapped to the screen).

Once we have read the swap chain features, we choose the surface format
to use. We prefer 32 bit BGRA colours in SRGB non-linear colour space.

TODO:
- Choose presentation mode
- Choose swap exent
- Create the swap chain
- Get swap chain images