WIP: Swap chain creation
Add functions for choosing present mode and swap extents

For choosing present mode, we check if the Mailbox present mode is
available. This mode pushes images as fast as possible, replacing older
images with newer images when the application is faster than the
display's refresh rate. If this mode is not available we use FIFO mode
which blocks while the queue is full.

For choosing swap extents, we will use the "current extent" of the
surface if it has been set by the window manager. Otherwise we use the
physical (pixel) size of the display as reported by the window manager -
clamped to be within the allowed range of image sizes for the surface as
reported by the surface capabilities.

Resources:
- https://vulkan-tutorial.com/Drawing_a_triangle/Presentation/Swap_chain
- https://khronos.org/registry/vulkan/specs/1.2-extensions/man/html/VkSurfaceCapabilitiesKHR.html
- https://docs.rs/winit/0.10.0/winit/struct.Window.html#method.get_inner_size

TODO:
- Create the swap chain
- Compare with vulkan-tutorial-rust