Complete swap chain creation
We combine the extents, surface format, and present mode that we
gathered previously and use that to define our swap chain create info
along with other paramters such as surface, image count, sharing mode, window
blending and more.

We create the swapchain by loading the swapchain function pointers and
calling create_swapchain. We add destruction of the swapchain to the
Drop implementation. We also fetch the swapchain images vector.

The swapchain function pointer, swapchain itself, swapchain images,
extent, and format are all stored on the application struct.

References:
- https://vulkan-tutorial.com/Drawing_a_triangle/Presentation/Swap_chain#page_Retrieving-the-swap-chain-images
- https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/VkSwapchainCreateInfoKHR.html