Frame buffers
Creates a frame buffer for each swap chain image and binds it to the
render pass. Each swap chain image is configured as the color image
attachment for the render pass in the frame buffer attachment.

They are destroyed before the image views when the application struct is
dropped.

Resources:
* https://vulkan-tutorial.com/Drawing_a_triangle/Drawing/Framebuffers