feat: Draw a quad using 4 vertices and an index buffer
We define index data to draw two triangles from 4 vertices. We create a
new buffer and memory to store the index buffer, following a similar
procedure as for the vertex buffer. Lastly, we change our commands so
that the vertex buffer gets bound, and we use DrawIndexed instead of
Draw.

We also clean up the buffers when the program exits.